pyproject.toml
[tool.poetry]
name = "withings-fastapi"
version = "0.1.0"
description = ""
authors = ["matsunagadaiki151"]
readme = "README.md"
packages = [{include = "withings_fastapi", from = "src"},
            {include = "utils"},
            {include = "server"}]


[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.17.0"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
requests = "^2.28.2"
python-dotenv = "^0.21.1"
types-requests = "^2.28.11.8"
flask = "^2.2.2"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.19.6"
virtualenv = "^20.17.1"
pre-commit = "^3.0.1"
plotly = "^5.13.0"
json-server = "^0.1.3"

[tool.poetry.scripts]
start = "server.main:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
mypy = "^0.991"
isort = "^5.12.0"
sphinx = "^6.1.3"
sphinx-pyproject = "^0.1.0"
flake8-isort = "^6.0.0"
flake8-bugbear = "^23.1.20"
flake8-builtins = "^2.1.0"
flake8-unused-arguments = "^0.0.12"
flake8-pytest-style = "^1.6.0"
pep8-naming = "^0.13.3"
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
# エラー時のメッセージを詳細表示
show_error_context = true
# エラー発生箇所の行数/列数を表示
show_column_numbers = true
# import 先のチェックを行わない (デフォルトだとサードパーティーライブラリまでチェックする)
ignore_missing_imports = true
# 関数定義の引数/戻り値に型アノテーション必須
disallow_untyped_defs = true
# デフォルト引数に None を取る場合型アノテーションに Optional 必須
no_implicit_optional = true
# 戻り値が Any 型ではない関数の戻り値の型アノテーションが Any のとき警告
warn_return_any = true
# mypy エラーに該当しない箇所に `# type: ignore` コメントが付与されていたら警告
# ※ `# type: ignore` が付与されている箇所は mypy のエラーを無視出来る
warn_unused_ignores = true
# 冗長なキャストに警告
warn_redundant_casts = true
